requirements.txt
django>=3.1,<3.2
djangorestframework>=3.12,<3.13
django-environ
psycopg2-binary
flake8
gunicorn
WhiteNoise